refactor : 상대주소 -> 절대주소(@사용)으로 변경
